+++ example/lib/data_sources.dart
@@ -148,9 +148,9 @@ class DessertDataSource extends DataTableSource {
       index: index,
       selected: dessert.selected,
       color: color != null
-          ? MaterialStateProperty.all(color)
+          ? WidgetStateProperty.all(color)
           : (hasZebraStripes && index.isEven
-              ? MaterialStateProperty.all(Theme.of(context).highlightColor)
+              ? WidgetStateProperty.all(Theme.of(context).highlightColor)
               : null),
       onSelectChanged: (value) {
         if (dessert.selected != value) {
@@ -212,7 +212,7 @@ class DessertDataSource extends DataTableSource {
 }
 
 /// Async datasource for AsynPaginatedDataTabke2 example. Based on AsyncDataTableSource which
-/// is an extension to FLutter's DataTableSource and aimed at solving
+/// is an extension to Flutter's DataTableSource and aimed at solving
 /// saync data fetching scenarious by paginated table (such as using Web API)
 class DessertDataSourceAsync extends AsyncDataTableSource {
   DessertDataSourceAsync() {
@@ -286,7 +286,7 @@ class DessertDataSourceAsync extends AsyncDataTableSource {
         x.data.map((dessert) {
           return DataRow(
             key: ValueKey<int>(dessert.id),
-            selected: dessert.selected,
+            //selected: dessert.selected,
             onSelectChanged: (value) {
               if (value != null) {
                 setRowSelection(ValueKey<int>(dessert.id), value);