+++ lib/src/data_table_2.dart
@@ -87,7 +87,7 @@ class DataRow2 extends DataRow {
     LocalKey? key,
     bool? selected,
     ValueChanged<bool?>? onSelectChanged,
-    MaterialStateProperty<Color?>? color,
+    WidgetStateProperty<Color?>? color,
     Decoration? decoration,
     List<DataCell>? cells,
     double? specificRowHeight,
@@ -166,6 +166,7 @@ class DataTable2 extends DataTable {
     this.checkboxAlignment = Alignment.center,
     this.bottomMargin,
     super.columnSpacing,
+    this.showHeadingCheckBox = true,
     super.showCheckboxColumn = true,
     super.showBottomBorder = false,
     super.dividerThickness,
@@ -249,6 +250,11 @@ class DataTable2 extends DataTable {
   /// Alignment of the checkbox if it is displayed
   /// Defaults to the [Alignment.center]
   final Alignment checkboxAlignment;
+
+  /// Whether to display heading checkbox or not if the checkbox column is present. Defaults to true.
+  /// Also check [DataTable.showCheckboxColumn] for details
+  /// on how to control the checkbox column visibility
+  final bool showHeadingCheckBox;
 
   /// Overrides theme of the checkbox that is displayed in the checkbox column
   /// in each data row (should checkboxes be enabled)
@@ -360,7 +366,7 @@ class DataTable2 extends DataTable {
       required bool? checked,
       required VoidCallback? onRowTap,
       required ValueChanged<bool?>? onCheckboxChanged,
-      required MaterialStateProperty<Color?>? overlayColor,
+      required WidgetStateProperty<Color?>? overlayColor,
       required CheckboxThemeData? checkboxTheme,
       required bool tristate,
       required double? rowHeight}) {
@@ -417,7 +423,7 @@ class DataTable2 extends DataTable {
       required bool sorted,
       required bool ascending,
       required double effectiveHeadingRowHeight,
-      required MaterialStateProperty<Color?>? overlayColor}) {
+      required WidgetStateProperty<Color?>? overlayColor}) {
     final ThemeData themeData = Theme.of(context);
 
     var customArrows =
@@ -489,7 +495,7 @@ class DataTable2 extends DataTable {
       required GestureTapCallback? onRowSecondaryTap,
       required GestureTapDownCallback? onRowSecondaryTapDown,
       required VoidCallback? onSelectChanged,
-      required MaterialStateProperty<Color?>? overlayColor}) {
+      required WidgetStateProperty<Color?>? overlayColor}) {
     final ThemeData themeData = Theme.of(context);
     final DataTableThemeData dataTableTheme = DataTableTheme.of(context);
 
@@ -587,9 +593,9 @@ class DataTable2 extends DataTable {
         headingRowColor ?? theme.dataTableTheme.headingRowColor;
     final effectiveDataRowColor =
         dataRowColor ?? theme.dataTableTheme.dataRowColor;
-    final defaultRowColor = MaterialStateProperty.resolveWith(
-      (Set<MaterialState> states) {
-        if (states.contains(MaterialState.selected)) {
+    final defaultRowColor = WidgetStateProperty.resolveWith(
+      (Set<WidgetState> states) {
+        if (states.contains(WidgetState.selected)) {
           return theme.colorScheme.primary.withOpacity(0.08);
         }
         return null;
@@ -657,13 +663,13 @@ class DataTable2 extends DataTable {
                         context,
                         theme,
                         fixedColumnsColor != null
-                            ? MaterialStatePropertyAll(fixedColumnsColor)
+                            ? WidgetStatePropertyAll(fixedColumnsColor)
                             : effectiveHeadingRowColor,
                         actualFixedColumns),
                     ..._buildTableRows(
                         anyRowSelectable,
                         fixedColumnsColor != null
-                            ? MaterialStatePropertyAll(fixedColumnsColor)
+                            ? WidgetStatePropertyAll(fixedColumnsColor)
                             : effectiveDataRowColor,
                         context,
                         theme,
@@ -677,7 +683,7 @@ class DataTable2 extends DataTable {
                 : _buildTableRows(
                     anyRowSelectable,
                     fixedColumnsColor != null
-                        ? MaterialStatePropertyAll(fixedColumnsColor)
+                        ? WidgetStatePropertyAll(fixedColumnsColor)
                         : effectiveDataRowColor,
                     context,
                     theme,
@@ -726,7 +732,7 @@ class DataTable2 extends DataTable {
                         context,
                         theme,
                         fixedCornerColor != null
-                            ? MaterialStatePropertyAll(fixedCornerColor)
+                            ? WidgetStatePropertyAll(fixedCornerColor)
                             : effectiveHeadingRowColor,
                         actualFixedColumns)
                   ]
@@ -735,13 +741,13 @@ class DataTable2 extends DataTable {
                         context,
                         theme,
                         fixedCornerColor != null
-                            ? MaterialStatePropertyAll(fixedCornerColor)
+                            ? WidgetStatePropertyAll(fixedCornerColor)
                             : effectiveHeadingRowColor,
                         actualFixedColumns),
                     ..._buildTableRows(
                         anyRowSelectable,
                         fixedCornerColor != null
-                            ? MaterialStatePropertyAll(fixedCornerColor)
+                            ? WidgetStatePropertyAll(fixedCornerColor)
                             : effectiveDataRowColor,
                         context,
                         theme,
@@ -1036,11 +1042,10 @@ class DataTable2 extends DataTable {
                   thumbVisibility: isHorizontalScrollBarVisible ??
                       (isiOS
                           ? scrollBarTheme.thumbVisibility
-                              ?.resolve({MaterialState.hovered})
+                              ?.resolve({WidgetState.hovered})
                           : null),
                   thickness: (isiOS
-                      ? scrollBarTheme.thickness
-                          ?.resolve({MaterialState.hovered})
+                      ? scrollBarTheme.thickness?.resolve({WidgetState.hovered})
                       : null),
                   controller: coreHorizontalController,
                   child: Column(mainAxisSize: MainAxisSize.min, children: [
@@ -1066,11 +1071,11 @@ class DataTable2 extends DataTable {
                             thumbVisibility: isVerticalScrollBarVisible ??
                                 (isiOS
                                     ? scrollBarTheme.thumbVisibility
-                                        ?.resolve({MaterialState.hovered})
+                                        ?.resolve({WidgetState.hovered})
                                     : null),
                             thickness: (isiOS
                                 ? scrollBarTheme.thickness
-                                    ?.resolve({MaterialState.hovered})
+                                    ?.resolve({WidgetState.hovered})
                                 : null),
                             controller: coreVerticalController,
                             child: SingleChildScrollView(
@@ -1158,7 +1163,7 @@ class DataTable2 extends DataTable {
       List<TableRow>? fixedColumnRows,
       List<DataRow> rows,
       int actualFixedRows,
-      MaterialStateProperty<Color?>? effectiveDataRowColor) {
+      WidgetStateProperty<Color?>? effectiveDataRowColor) {
     double checkBoxWidth = 0;
 
     if (displayCheckboxColumn) {
@@ -1168,16 +1173,21 @@ class DataTable2 extends DataTable {
       tableColumns[0] = FixedColumnWidth(checkBoxWidth);
 
       // Create heading twice, in the heading row used as back-up for the case of no data and any of the xxx_rows table
-      headingRow.children[0] = _buildCheckbox(
-          context: context,
-          checked: someChecked ? null : allChecked,
-          onRowTap: null,
-          onCheckboxChanged: (bool? checked) =>
-              _handleSelectAll(checked, someChecked),
-          overlayColor: null,
-          checkboxTheme: headingCheckboxTheme,
-          tristate: true,
-          rowHeight: headingHeight);
+
+      headingRow.children[0] = showHeadingCheckBox
+          ? _buildCheckbox(
+              context: context,
+              checked: someChecked ? null : allChecked,
+              onRowTap: null,
+              onCheckboxChanged: (bool? checked) =>
+                  _handleSelectAll(checked, someChecked),
+              overlayColor: null,
+              checkboxTheme: headingCheckboxTheme,
+              tristate: true,
+              rowHeight: headingHeight)
+          : SizedBox(
+              height: headingHeight,
+            );
 
       if (fixedCornerRows != null) {
         fixedCornerRows[0].children[0] = headingRow.children[0];
@@ -1321,11 +1331,11 @@ class DataTable2 extends DataTable {
 
   List<TableRow> _buildTableRows(
       bool anyRowSelectable,
-      MaterialStateProperty<Color?>? effectiveDataRowColor,
+      WidgetStateProperty<Color?>? effectiveDataRowColor,
       BuildContext context,
       ThemeData theme,
       int numberOfCols,
-      MaterialStateProperty<Color?> defaultRowColor,
+      WidgetStateProperty<Color?> defaultRowColor,
       TableRow? headingRow,
       [int skipRows = 0,
       int takeRows = 0,
@@ -1342,9 +1352,9 @@ class DataTable2 extends DataTable {
           final bool isSelected = rows[rowStartIndex + actualIndex].selected;
           final bool isDisabled = anyRowSelectable &&
               rows[rowStartIndex + actualIndex].onSelectChanged == null;
-          final Set<MaterialState> states = <MaterialState>{
-            if (isSelected) MaterialState.selected,
-            if (isDisabled) MaterialState.disabled,
+          final Set<WidgetState> states = <WidgetState>{
+            if (isSelected) WidgetState.selected,
+            if (isDisabled) WidgetState.disabled,
           };
           final Color? resolvedDataRowColor = (forceEffectiveDataRowColor
                   ? effectiveDataRowColor
@@ -1390,11 +1400,8 @@ class DataTable2 extends DataTable {
     return tableRows;
   }
 
-  TableRow _buildHeadingRow(
-      BuildContext context,
-      ThemeData theme,
-      MaterialStateProperty<Color?>? effectiveHeadingRowColor,
-      int numberOfCols) {
+  TableRow _buildHeadingRow(BuildContext context, ThemeData theme,
+      WidgetStateProperty<Color?>? effectiveHeadingRowColor, int numberOfCols) {
     var headingRow = TableRow(
       key: _headingRowKey,
       decoration: BoxDecoration(
@@ -1411,7 +1418,7 @@ class DataTable2 extends DataTable {
                     _dividerThickness,
               ))
             : null,
-        color: effectiveHeadingRowColor?.resolve(<MaterialState>{}),
+        color: effectiveHeadingRowColor?.resolve(<WidgetState>{}),
       ).copyWith(
         color: headingRowDecoration?.color,
         image: headingRowDecoration?.image,