+++ test/custom_features_test.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: avoid_print
 
 @TestOn('!chrome')
+library;
 
 import 'package:data_table_2/data_table_2.dart';
 import 'package:flutter/material.dart';
@@ -58,7 +59,7 @@ void main() {
           Theme(
               data: ThemeData(
                   scrollbarTheme: ScrollbarThemeData(
-                      thumbVisibility: MaterialStateProperty.all(true))),
+                      thumbVisibility: WidgetStateProperty.all(true))),
               child: buildTable()),
           const Size(250, 300));
 
@@ -661,11 +662,10 @@ void main() {
       await wrapWidgetSetSurf(
           tester,
           buildPaginatedTable(
-              headingRowColor: MaterialStateProperty.all(Colors.yellow)));
+              headingRowColor: WidgetStateProperty.all(Colors.yellow)));
 
       var t = tester.widget(find.byType(DataTable2)) as DataTable2;
-      expect(
-          t.headingRowColor!.resolve({MaterialState.focused}), Colors.yellow);
+      expect(t.headingRowColor!.resolve({WidgetState.focused}), Colors.yellow);
 
       await wrapWidgetSetSurf(tester, buildPaginatedTable());
 
@@ -1921,9 +1921,9 @@ void main() {
     Widget buildTable() {
       return DataTable2(
         headingCheckboxTheme:
-            CheckboxThemeData(fillColor: MaterialStateProperty.all(Colors.red)),
+            CheckboxThemeData(fillColor: WidgetStateProperty.all(Colors.red)),
         datarowCheckboxTheme: CheckboxThemeData(
-            fillColor: MaterialStateProperty.all(Colors.yellow)),
+            fillColor: WidgetStateProperty.all(Colors.yellow)),
         showCheckboxColumn: true,
         onSelectAll: (value) {},
         columns: <DataColumn>[
@@ -1965,7 +1965,7 @@ void main() {
                 .data
                 .checkboxTheme
                 .fillColor!
-                .resolve({MaterialState.hovered}) ==
+                .resolve({WidgetState.hovered}) ==
             Colors.red);
     expect(h.length, 1);
     expect((h.first.widget as Theme).child.runtimeType, Checkbox);
@@ -1976,10 +1976,47 @@ void main() {
                 .data
                 .checkboxTheme
                 .fillColor!
-                .resolve({MaterialState.hovered}) ==
+                .resolve({WidgetState.hovered}) ==
             Colors.yellow);
     expect(d.length, kDesserts.length);
     expect((d.first.widget as Theme).child.runtimeType, Checkbox);
+  });
+
+  testWidgets('DataTable2 heading checkbox', (WidgetTester tester) async {
+    final List<DataColumn2> columns = [
+      const DataColumn2(label: Text('Column1')),
+    ];
+    final List<DataRow2> rows = [
+      DataRow2(
+        onSelectChanged: (value) {},
+        cells: const [DataCell(Text('Content1'))],
+      ),
+    ];
+
+    // Check if heading checkbox is shown by default
+    await wrapWidgetSetSurf(
+        tester,
+        buildTable(
+          columns: columns,
+          rows: rows,
+        ));
+
+    await tester.pumpAndSettle();
+    expect(find.text('Column1'), findsOneWidget);
+    expect(find.byType(Checkbox), findsNWidgets(2));
+
+    // Check if heading checkbox is hided
+    await wrapWidgetSetSurf(
+        tester,
+        buildTable(
+          showHeadingCheckbox: false,
+          columns: columns,
+          rows: rows,
+        ));
+
+    await tester.pumpAndSettle();
+    expect(find.text('Column1'), findsOneWidget);
+    expect(find.byType(Checkbox), findsOneWidget);
   });
 }