+++ CHANGELOG.md
@@ -1,5 +1,18 @@
+## 2.5.15
+- `dart fromat .`
+
+## 2.5.14
+- Fixing linter warnings (replacing deprecated MaterialStateProperty)
+
+## 2.5.12
+- Add option to hide heading checkbox in data table PR#270
+
+## 2.5.11
+- Fixed Async example (range selector)
+- Added `isHorizontalScrollBarVisible` and `isVerticalScrollBarVisible` to `PaginatedDataTable` and `AsyncPaginatedDataTable`
+
 ## 2.5.10
-- Added [DataTable2.decoratio]
+- Added [DataTable2.decoration]
 - Example with rows styles as rounded corners rectangles
 
 ## 2.5.9